Make G-Machine transpiler more Haddock-friendly
