+++ src/Minicute/Transpilers/GMachine.hs
@@ -1,221 +1,15 @@
-{-# LANGUAGE OverloadedStrings #-}
+{-# OPTIONS_HADDOCK not-home #-}
+{-|
+The transpiler from a miniCUTE program to a G-Machine program
+-}
 module Minicute.Transpilers.GMachine
-  ( -- * Transpilers
-    -- ** Calling Convention
-    -- $callingConvention
+  ( module Minicute.Types.GMachine.Instruction
+  , module Minicute.Types.Minicute.Program
+  , transpileProgram
 
-    -- ** Transpiler Functions
-    -- *** Transpilers for a Program
-    transpileProgram
-  , transpileSc
-
-    -- *** Transpilers for an Expression
-  , transpileRE
-  , transpileSE
-  , transpileNE
-  , transpilePE
-
-    -- *** Transpilers for others
-  , transpileLet
-  , transpileMatch
-
-    -- ** Types
-  , TranspilerE
-  , TranspilerEEnv
-
-    -- * Initial Instructions
   , initialCode
   ) where
 
-import Control.Lens.Each
-import Control.Lens.Operators
-import Control.Lens.Wrapped ( _Wrapped )
-import Data.List
+import Minicute.Internal.Transpilers.GMachine
 import Minicute.Types.GMachine.Instruction
-import Minicute.Types.Minicute.Precedence
 import Minicute.Types.Minicute.Program
-
-import qualified Data.Map as Map
-
-{- $callingConvention
-Which calling convention we gonna use?
-
-1. Try simple one
-
-    1. caller
-
-        - Push arguments in caller
-        - Evaluate result in caller
-
-    1. callee
-
-        - Update results in callee
-        - Pop arguments in callee
--}
-
-
-initialCode ::[Instruction]
-initialCode = [IMakeGlobal "main", IEval]
-
-
-{-|
-Transpiler for a __Program__.
--}
-transpileProgram :: MainProgram -> GMachineProgram
-transpileProgram program = program ^. _Wrapped <&> transpileSc
-
-{-|
-Transpiler for a __S__uper__c__ombinator (top-level function definition).
--}
-transpileSc :: MainSupercombinator -> GMachineSupercombinator
-transpileSc sc = (scBinder, scArgsLength, scInsts)
-  where
-    scBinder = sc ^. _supercombinatorBinder
-
-    scInsts = sc ^. _supercombinatorBody & transpileRE scArgsEnv
-
-    scArgsEnv = Map.fromList $ zip scArgs [0..]
-    scArgsLength = length scArgs
-    scArgs = sc ^. _supercombinatorArguments
-
-type TranspilerE a = TranspilerEEnv -> a -> GMachineExpression
-
-{-|
-Transpiler for a __R__oot __E__xpression.
--}
-transpileRE :: TranspilerE MainExpression
-transpileRE env (EInteger n) = [IPushBasicValue n, IUpdateAsInteger (getEnvSize env), IReturn]
-transpileRE env (EConstructor tag 0) = [IPushBasicValue tag, IUpdateAsConstructor (getEnvSize env), IReturn]
-transpileRE env e@(EApplication2 (EVariableIdentifier op) _ _)
-  | Just _ <- lookup op binaryIntegerPrecendenceTable
-  = transpilePE env e <> [IUpdateAsInteger (getEnvSize env), IReturn]
-  | Just _ <- lookup op binaryDataPrecendenceTable
-  = transpilePE env e <> [IUpdateAsConstructor (getEnvSize env), IReturn]
-transpileRE env (ELet flag lDefs body) = transpileLet transpileRE env (flag, lDefs, body)
-transpileRE env (EMatch body mCases) = transpileMatch transpileSE (const transpileRE) env (body, mCases)
--- Should following really use a strict expression?
-transpileRE env e = transpileSE env e <> [IUpdate envSize1, IPop envSize1, IUnwind]
-  where
-    envSize1 = getEnvSize1 env
-
-{-|
-Transpiler for a __S__trict __E__xpression.
-All instruction generated by this transpiler should put the address of a whnf node on the top of the stack.
-
-This transpiler is not able to directly transpile variable expressions because value of variable can be a non-whnf node.
--}
-transpileSE :: TranspilerE MainExpression
-transpileSE _ (EInteger n) = [IMakeInteger n]
-transpileSE _ (EConstructor tag arity) = [IMakeConstructor tag arity]
-transpileSE env e@(EApplication2 (EVariableIdentifier op) _ _)
-  | Just _ <- lookup op binaryIntegerPrecendenceTable
-  = transpilePE env e <> [IWrapAsInteger]
-  | Just _ <- lookup op binaryDataPrecendenceTable
-  = transpilePE env e <> [IWrapAsConstructor]
-transpileSE env e = transpileNE env e <> [IEval]
-
-{-|
-Transpiler for a __P__rimitive __E__xpression.
-All instruction generated by this transpiler should put a primitive value on the top of the value stack.
--}
-transpilePE :: TranspilerE MainExpression
-transpilePE _ (EInteger n) = [IPushBasicValue n]
-transpilePE _ (EConstructor tag 0) = [IPushBasicValue tag]
-transpilePE env (EApplication2 (EVariableIdentifier op) e1 e2)
-  | Just _ <- lookup op binaryPrecedenceTable
-  = transpilePE env e1 <> transpilePE env e2 <> [IPrimitive (getPrimitiveOperator op)]
-transpilePE env (ELet flag lDefs body) = transpileLet transpilePE env (flag, lDefs, body) <> [IPop (length lDefs)]
-transpilePE env e = transpileSE env e <> [IPushExtractedValue]
-
-{-|
-Transpiler for a __N__on-strict __E__xpression.
-All instruction generated by this transpiler should put the address of a node on the top of the stack.
--}
-transpileNE :: TranspilerE MainExpression
-transpileNE _ (EInteger n) = [IMakeInteger n]
-transpileNE _ (EConstructor tag arity) = [IMakeConstructor tag arity]
-transpileNE env (EVariable v)
-  | Just index <- Map.lookup v env = [ICopyArgument index]
-  | otherwise = [IMakeGlobal v]
-transpileNE env (EApplication e1 e2)
-  = transpileNE env e1 <> transpileNE (addEnvOffset1 env) e2 <> [IMakeApplication]
-transpileNE env (ELet flag lDefs body) = transpileLet transpileNE env (flag, lDefs, body) <> [IDig (length lDefs)]
-transpileNE _ _ = error "Not yet implemented"
-
-{-|
-Transpiler for a __Let__ expression.
--}
-transpileLet :: TranspilerE MainExpression -> TranspilerE (IsRecursive, [MainLetDefinition], MainExpression)
-transpileLet transpileBody env (flag, lDefs, body)
-  | isRecursive flag = transpileLetRecDefs env' lDefs <> bodyInst
-  | otherwise = transpileLetDefs env lDefs <> bodyInst
-  where
-    bodyInst = transpileBody env' body
-    env' = updateLetEnv lDefs env
-
-transpileLetDefs :: TranspilerE [MainLetDefinition]
-transpileLetDefs env lDefs
-  = concat (zipWith transpileNE envs lDefsBodies)
-  where
-    envs = iterate addEnvOffset1 env
-    lDefsBodies = lDefs ^.. each . _letDefinitionBody
-
-transpileLetRecDefs :: TranspilerE [MainLetDefinition]
-transpileLetRecDefs env lDefs
-  = [IMakePlaceholders len] <> concatMap ((<> [IUpdate len]) . transpileNE env) lDefsBodies
-  where
-    lDefsBodies = lDefs ^.. each . _letDefinitionBody
-    len = length lDefs
-
-updateLetEnv :: [MainLetDefinition] -> TranspilerEEnv -> TranspilerEEnv
-updateLetEnv lDefs env = envOfLDefs <> addEnvOffset len env
-  where
-    envOfLDefs = Map.fromList (zip lDefsBinders [len - 1, len - 2 .. 0])
-    lDefsBinders = lDefs ^.. each . _letDefinitionBinder
-    len = length lDefs
-
-{-|
-Transpiler for a __Match__ expression.
--}
-transpileMatch :: TranspilerE MainExpression -> (Integer -> TranspilerE MainExpression) -> TranspilerE (MainExpression, [MainMatchCase])
-transpileMatch transpileBody transpileCase env (body, mCases) = bodyInst <> [IMatch (makeMatchTable transpileCase env mCases)]
-  where
-    bodyInst = transpileBody env body
-
-makeMatchTable :: (Integer -> TranspilerE MainExpression) -> TranspilerEEnv -> [MainMatchCase] -> MatchTable
-makeMatchTable transpileCase env = MatchTable . fmap (makeMatchEntry transpileCase env)
-
-makeMatchEntry :: (Integer -> TranspilerE MainExpression) -> TranspilerEEnv -> MainMatchCase -> MatchEntry
-makeMatchEntry transpileCase env mCase = MatchEntry (caseTag, caseInst)
-  where
-    caseInst = [IDestruct caseArgsLen] <> transpileCase caseArgsLen env' caseBody
-    caseBody = mCase ^. _matchCaseBody
-    caseTag = mCase ^. _matchCaseTag
-
-    env' = Map.fromList (zip caseArgs [0..]) <> addEnvOffset (fromInteger caseArgsLen) env
-
-    caseArgsLen = genericLength caseArgs
-    caseArgs = mCase ^. _matchCaseArguments
-
-
-type TranspilerEEnv = Map.Map Identifier Int
-
-getEnvSize1 :: TranspilerEEnv -> Int
-getEnvSize1 = (+ 1) . getEnvSize
-
-getEnvSize :: TranspilerEEnv -> Int
-getEnvSize = Map.size
-
-addEnvOffset1 :: TranspilerEEnv -> TranspilerEEnv
-addEnvOffset1 = Map.map (+ 1)
-
-addEnvOffset :: Int -> TranspilerEEnv -> TranspilerEEnv
-addEnvOffset n = Map.map (+ n)
-
-
-getPrimitiveOperator :: String -> PrimitiveOperator
-getPrimitiveOperator "+" = POAdd
-getPrimitiveOperator "-" = POSub
-getPrimitiveOperator "*" = POMul
-getPrimitiveOperator "/" = PODiv
-getPrimitiveOperator _ = error "Not implemented yet"
